requirements.txt
fastapi
uvicorn
PyJWT[crypto]
uvicorn
python-multipart
cryptography
pycouchdb